feat:echo functionallity completed

+++ cli/src/models/commands/echo_response.rs
@@ -3,6 +3,6 @@ use serde::Serialize;
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct EchoResponse {
-    pub code: u16,
+    pub code: i64,
     pub message: String,
 }
